Enable AMP autocast fp16 in patcher and TinyLM training

+++ scripts/train_tiny.py
@@ -154,7 +154,7 @@ def main():
             for group in optimizer.param_groups:
                 group["lr"] = lr
 
-            with torch.cuda.amp.autocast(enabled=(device.type == "cuda")):
+            with torch.cuda.amp.autocast(enabled=(device.type == "cuda"), dtype=torch.float16):
                 _, loss = model(x, y)
                 loss = loss / grad_accum_steps
